jit: Be very clear on sign extension.

+++ Core/MIPS/MIPSCodeUtils.cpp
@@ -30,7 +30,7 @@ namespace MIPSCodeUtils
 #define _RT   ((op>>16) & 0x1F)
 #define _IMM16 (signed short)(op & 0xFFFF)
 #define _IMM26 (op & 0x03FFFFFF)
-#define TARGET16 ((int)((uint32_t)_IMM16 << 2))
+#define TARGET16 ((int)((uint32_t)(int)_IMM16 << 2))
 #define TARGET26 (_IMM26 << 2)
 
 	u32 GetJumpTarget(u32 addr) {
